smb2: finish decomposing smb_server.c into per-concern translation units (#75)
Move smb_ioctl (and its SMB2_FSCTL_PIPE_TRANSCEIVE/VALIDATE_NEGOTIATE_INFO/
DFS_GET_REFERRALS handling) into a new smb_server_srvsvc.c, promoting it to
external linkage per its slot in the handler table. Remove the 23 duplicate
VFS-handler function definitions left behind in smb_server.c by the prior
split into smb_server_vfs.c, keeping smb_normalize_share_root (which is not
a VFS handler) in place. Wire the four new translation units into the
Makefile alongside smb_server.c.

Pure code motion: every relocated function body is byte-identical to the
pre-split smb_server.c modulo a leading `static` removal where the symbol
is now referenced across translation units. No logic changes.
